requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
transformers==4.36.2
beautifulsoup4==4.12.2
scholarly==1.7.11
arxiv==1.4.8
google-generativeai==0.3.1
numpy==1.26.2
gunicorn==21.2.0
lxml==4.9.3 